finished

+++ app/src/file.py
@@ -4,8 +4,8 @@ THETA_FILE_PATH = os.path.join(os.path.dirname(os.path.realpath(__file__))[:-3],
 def read_theta_from_file(path=THETA_FILE_PATH):
 	try:
 		with open(path, 'r') as f:
-			theta0 = int(f.readline())
-			theta1 = int(f.readline())
+			theta0 = float(f.readline())
+			theta1 = float(f.readline())
 	except:
 		theta0, theta1 = 0, 0
 	return theta0, theta1
